Put proof of concept into proper class form
